Catch error on fwrite, and display a friendlier message when submitting changes to prefs/sigs.  Thanks Simon Dick for the suggestion

+++ squirrelmail/functions/file_prefs.php
@@ -97,7 +97,7 @@ function savePrefValues($data_dir, $username) {
     $filename = getHashedFile($username, $data_dir, "$username.pref");
 
     /* Open the file for writing, or else display an error to the user. */
-    if(!$file = @fopen("$filename.tmp", 'w')) {
+    if(!$file = @fopen($filename. '.tmp', 'w')) {
         include_once( '../functions/display_messages.php' );
         logout_error( sprintf( _("Preference file, %s, could not be opened. Contact your system administrator to resolve this issue."), "$filename.tmp") );
         exit;
@@ -105,12 +105,16 @@ function savePrefValues($data_dir, $username) {
 
     foreach ($prefs_cache as $Key => $Value) {
         if (isset($Value)) {
-            fwrite($file, $Key . '=' . $Value . "\n");
+            $tmpwrite = @fwrite($file, $Key . '=' . $Value . "\n");
+            if ($tmpwrite == -1) {
+               logout_error( sprintf( _("Preference file, %s, could not be written. Contact your system administrator to resolve this issue.") , $filename . '.tmp') );
+               exit;
+            }
         }
     }
     fclose($file);
-    @copy("$filename.tmp","$filename");
-    @unlink("$filename.tmp");
+    @copy($filename . '.tmp',$filename);
+    @unlink($filename . '.tmp');
     chmod($filename, 0600);
 }
 
@@ -205,7 +209,12 @@ function setSig($data_dir, $username, $number, $value) {
         logout_error( sprintf( _("Signature file, %s, could not be opened. Contact your system administrator to resolve this issue."), "$filename.tmp") );
         exit;
     }
-    fwrite($file, $value);
+    $tmpwrite = @fwrite($file, $value);
+    if ($tmpwrite == -1) {
+       include_once( '../functions/display_messages.php' );
+       logout_error( sprintf( _("Signaute file, %s, could not be written. Contact your system administrator to resolve this issue.") , $filename . '.tmp'));
+       exit;
+    }
     fclose($file);
     copy("$filename.tmp","$filename");
     unlink("$filename.tmp");
